Lidar data computing

+++ lidar_dev/Core/Src/main.c
@@ -57,6 +57,7 @@ void SystemClock_Config(void);
 void HAL_UART_RxCpltCallback(UART_HandleTypeDef * huart){
 	if(huart->Instance == USART1){
 		ydlidar_x4_irq_cb(&h_ylidar_x4);
+		//ydlidar_x4_sort_smpl(&h_ylidar_x4);
 		//HAL_GPIO_WritePin(LD2_GPIO_Port, LD2_Pin, 1);
 		//ydlidar_x4_irq_cb(h_ylidar_x4);
 	}
@@ -145,7 +146,7 @@ int main(void)
 		for(int i=0;i<600;i++){
 			if(h_ylidar_x4.rev_smpls[i][1] != 0){
 				uint16_t size = snprintf(mystring,50,"angle = %d, dist = %d\r\n",h_ylidar_x4.rev_smpls[i][0],h_ylidar_x4.rev_smpls[i][1]);
-				HAL_UART_Transmit(&huart2, mystring, size, HAL_MAX_DELAY);
+				HAL_UART_Transmit(&huart2, (uint8_t) mystring, size, HAL_MAX_DELAY);
 			}
 		}
     /* USER CODE END WHILE */
